feat: implement get endpoint and refactor repository and service class
