+++ chat/src/main/kotlin/net/thechance/chat/api/controller/ContactController.kt
@@ -1,15 +1,18 @@
 package net.thechance.chat.api.controller
 
-import chat.api.dto.ContactResponse
-import chat.api.dto.PagedResponse
-import chat.api.dto.toResponse
+import net.thechance.chat.api.dto.ContactResponse
+import net.thechance.chat.api.dto.PagedResponse
+import net.thechance.chat.api.dto.toResponse
 import chat.service.toResponse
 import net.thechance.chat.service.ContactService
 import org.springframework.data.domain.Pageable
 import org.springframework.http.ResponseEntity
+import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
+import java.security.Principal
+import java.util.UUID
 
 @RestController
 @RequestMapping("/contacts")
@@ -18,8 +21,11 @@ class ContactController(
 ) {
 
     @GetMapping
-    fun getPagedContact(pageable: Pageable): ResponseEntity<PagedResponse<ContactResponse>> {
-        val page = contactService.getPagedContact(pageable).map { it.toResponse() }
+    fun getPagedContact(
+        pageable: Pageable,
+        @AuthenticationPrincipal userId: UUID
+    ): ResponseEntity<PagedResponse<ContactResponse>> {
+        val page = contactService.getPagedContactByUserId(userId, pageable).map { it.toResponse() }
         return ResponseEntity.ok(page.toResponse())
     }
 }
